lfd: reorganize into daemon, execution, and runs packages
Split monolithic modules into focused packages:
- daemon/: server, client, manager (née scheduler), launchd, protocol
- execution/: runner, worker, collector
- runs/: loop, run, schedule, session, subscription persistence
