+++ src/loopflow/lfd/daemon/manager.py
@@ -1,4 +1,8 @@
-"""Scheduler for coordinating parallel loop execution."""
+"""Manager for coordinating parallel worker execution.
+
+The manager controls how many workers can run simultaneously and
+tracks global PR limits across all triggers.
+"""
 
 import threading
 from dataclasses import dataclass
@@ -7,41 +11,41 @@ from pathlib import Path
 import yaml
 
 from loopflow.lfd.db import list_all_triggers
-from loopflow.lfd.loops import count_outstanding
+from loopflow.lfd.runs.loop import count_outstanding
 
 
 @dataclass
-class SchedulerConfig:
-    """Configuration for the loop scheduler."""
+class ManagerConfig:
+    """Configuration for the worker manager."""
 
-    concurrency: int = 3  # Max parallel iterations
-    global_pr_limit: int = 15  # Total outstanding across all loops
+    concurrency: int = 3  # Max parallel workers
+    global_pr_limit: int = 15  # Total outstanding across all triggers
 
 
-def load_scheduler_config() -> SchedulerConfig:
-    """Load scheduler config from ~/.lf/lfd.yaml."""
+def load_manager_config() -> ManagerConfig:
+    """Load manager config from ~/.lf/lfd.yaml."""
     config_path = Path.home() / ".lf" / "lfd.yaml"
     if not config_path.exists():
-        return SchedulerConfig()
+        return ManagerConfig()
 
     try:
         data = yaml.safe_load(config_path.read_text()) or {}
-        return SchedulerConfig(
+        return ManagerConfig(
             concurrency=data.get("concurrency", 3),
             global_pr_limit=data.get("global_pr_limit", 15),
         )
     except Exception:
-        return SchedulerConfig()
+        return ManagerConfig()
 
 
-class Scheduler:
-    """Coordinates multiple loops with shared resource limits.
+class Manager:
+    """Coordinates multiple workers with shared resource limits.
 
     Thread-safe slot management for parallel iteration execution.
     """
 
-    def __init__(self, config: SchedulerConfig | None = None):
-        self.config = config or load_scheduler_config()
+    def __init__(self, config: ManagerConfig | None = None):
+        self.config = config or load_manager_config()
         self._running: set[str] = set()  # run IDs currently executing
         self._lock = threading.Lock()
 
@@ -71,7 +75,7 @@ class Scheduler:
         return total
 
     def can_start(self) -> tuple[bool, str | None]:
-        """Check if a new iteration can start.
+        """Check if a new worker can start.
 
         Returns (can_start, reason) where reason explains why if can't start.
         """
@@ -85,7 +89,7 @@ class Scheduler:
         return True, None
 
     def acquire(self, run_id: str) -> tuple[bool, str | None]:
-        """Try to acquire a slot for an iteration.
+        """Try to acquire a slot for a worker.
 
         Returns (acquired, reason) where reason explains why if not acquired.
         """
@@ -103,12 +107,12 @@ class Scheduler:
             return True, None
 
     def release(self, run_id: str) -> None:
-        """Release a slot when iteration completes."""
+        """Release a slot when worker completes."""
         with self._lock:
             self._running.discard(run_id)
 
     def get_status(self) -> dict:
-        """Get scheduler status for display."""
+        """Get manager status for display."""
         with self._lock:
             slots_used = len(self._running)
             running_ids = list(self._running)
@@ -122,19 +126,27 @@ class Scheduler:
         }
 
 
-# Global scheduler instance for daemon
-_scheduler: Scheduler | None = None
+# Global manager instance for daemon
+_manager: Manager | None = None
 
 
-def get_scheduler() -> Scheduler:
-    """Get or create the global scheduler instance."""
-    global _scheduler
-    if _scheduler is None:
-        _scheduler = Scheduler()
-    return _scheduler
+def get_manager() -> Manager:
+    """Get or create the global manager instance."""
+    global _manager
+    if _manager is None:
+        _manager = Manager()
+    return _manager
 
 
-def reset_scheduler() -> None:
-    """Reset the global scheduler (for testing)."""
-    global _scheduler
-    _scheduler = None
+def reset_manager() -> None:
+    """Reset the global manager (for testing)."""
+    global _manager
+    _manager = None
+
+
+# Backwards compatibility aliases
+Scheduler = Manager
+SchedulerConfig = ManagerConfig
+load_scheduler_config = load_manager_config
+get_scheduler = get_manager
+reset_scheduler = reset_manager
