+++ src/loopflow/lfd/execution/__init__.py
@@ -1,0 +1,7 @@
+"""Execution layer for lfd.
+
+This package handles actually running iterations:
+- runner.py: run_iteration - executes a single iteration of any trigger type
+- worker.py: run_loop_iterations - continuous execution for loops
+- collector.py: collects and streams output during execution
+"""
