+++ src/loopflow/lfd/execution/runner.py
@@ -1,19 +1,13 @@
-"""Loop iteration runner for lfd.
+"""Core iteration runner for lfd.
 
-Runs iterations of a loop until the PR limit is reached or an error occurs.
-Can be invoked directly as a subprocess for background execution.
-
-When running in background, coordinates with the daemon scheduler to respect
-global concurrency and PR limits.
+Executes a single iteration of any trigger type (Loop, Subscription, Schedule).
 """
 
 import json
-import socket
 import subprocess
 import sys
-import time
 import uuid
-from dataclasses import dataclass, replace
+from dataclasses import replace
 from datetime import datetime
 from pathlib import Path
 
@@ -40,158 +34,21 @@ from loopflow.lf.messages import generate_pr_message
 from loopflow.lf.worktrees import WorktreeError
 from loopflow.lf.worktrees import create as create_worktree
 from loopflow.lf.worktrees import remove as remove_worktree
-from loopflow.lfd.client import notify_event
-from loopflow.lfd.db import (
-    get_loop,
+from loopflow.lfd.daemon.client import notify_event
+from loopflow.lfd.models import Loop, Run, RunStatus
+from loopflow.lfd.runs.run import (
     save_run,
-    update_loop_iteration,
-    update_loop_pid,
-    update_loop_status,
     update_run_pr,
     update_run_status,
     update_run_step,
 )
-from loopflow.lfd.loops import count_outstanding
-from loopflow.lfd.models import Loop, Run, RunStatus, TriggerStatus
-
-SOCKET_PATH = Path.home() / ".lf" / "lfd.sock"
-SCHEDULER_POLL_INTERVAL = 30  # seconds between slot checks
-
-
-@dataclass
-class _VariantResult:
-    label: str
-    worktree: Path
-    exit_code: int
-    session_id: str
 
 
 def _iteration_branch_prefix(main_branch: str) -> str:
-    """Derive iteration branch prefix from main branch.
-
-    'product-engineer-main' → 'product-engineer'
-    'product-engineer-1-main' → 'product-engineer-1'
-    """
+    """Derive iteration branch prefix from main branch."""
     if main_branch.endswith("-main"):
         return main_branch[:-5]
     return main_branch
-
-
-def _scheduler_call(method: str, params: dict | None = None) -> dict | None:
-    """Make a synchronous call to the daemon scheduler.
-
-    Returns the result dict on success, None on connection failure.
-    """
-    if not SOCKET_PATH.exists():
-        return None
-
-    try:
-        sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-        sock.settimeout(5.0)
-        sock.connect(str(SOCKET_PATH))
-
-        request = {"method": method}
-        if params:
-            request["params"] = params
-
-        sock.sendall((json.dumps(request) + "\n").encode())
-
-        response_data = b""
-        while b"\n" not in response_data:
-            chunk = sock.recv(1024)
-            if not chunk:
-                break
-            response_data += chunk
-
-        sock.close()
-
-        if response_data:
-            response = json.loads(response_data.decode().strip())
-            if response.get("ok"):
-                return response.get("result", {})
-        return None
-    except Exception:
-        return None
-
-
-def _scheduler_acquire(run_id: str) -> tuple[bool, str | None]:
-    """Try to acquire a scheduler slot.
-
-    Returns (acquired, reason) when the daemon is available.
-    """
-    result = _scheduler_call("scheduler.acquire", {"run_id": run_id})
-    if result is None:
-        # Daemon not running, allow iteration (standalone mode)
-        return True, None
-    return result.get("acquired", False), result.get("reason")
-
-
-def _scheduler_release(run_id: str) -> None:
-    """Release a scheduler slot."""
-    _scheduler_call("scheduler.release", {"run_id": run_id})
-
-
-def run_loop_iterations(loop: Loop) -> None:
-    """Run loop iterations until PR limit is reached or error occurs."""
-    while True:
-        # Check if we should pause (per-loop limit)
-        outstanding = count_outstanding(loop)
-        if outstanding >= loop.pr_limit:
-            update_loop_status(loop.id, TriggerStatus.WAITING)
-            notify_event(
-                "loop.waiting",
-                {
-                    "loop_id": loop.id,
-                    "area": loop.area,
-                    "outstanding": outstanding,
-                    "limit": loop.pr_limit,
-                },
-            )
-            break
-
-        # Run one iteration
-        iteration = loop.iteration + 1
-        run_id = str(uuid.uuid4())
-
-        # Wait for scheduler slot (global concurrency)
-        while True:
-            acquired, reason = _scheduler_acquire(run_id)
-            if acquired:
-                break
-            notify_event(
-                "scheduler.waiting",
-                {
-                    "loop_id": loop.id,
-                    "area": loop.area,
-                    "reason": reason or "concurrency",
-                },
-            )
-            time.sleep(SCHEDULER_POLL_INTERVAL)
-
-        try:
-            success = run_iteration(loop, iteration, run_id)
-            if success:
-                update_loop_iteration(loop.id, iteration)
-                loop.iteration = iteration
-            else:
-                update_loop_status(loop.id, TriggerStatus.ERROR)
-                break
-        except Exception as e:
-            notify_event(
-                "loop.error",
-                {
-                    "loop_id": loop.id,
-                    "area": loop.area,
-                    "error": str(e),
-                },
-            )
-            update_loop_status(loop.id, TriggerStatus.ERROR)
-            break
-        finally:
-            _scheduler_release(run_id)
-
-    # Clear pid when done
-    update_loop_pid(loop.id, None)
 
 
 def _build_loop_prompt(
@@ -254,7 +111,7 @@ def _run_collector_step(
     collector_cmd = [
         sys.executable,
         "-m",
-        "loopflow.lfd.collector",
+        "loopflow.lfd.execution.collector",
         "--session-id",
         session_id,
         "--step",
@@ -279,6 +136,14 @@ def _run_collector_step(
         pass
 
     return result_code
+
+
+class _VariantResult:
+    def __init__(self, label: str, worktree: Path, exit_code: int, session_id: str):
+        self.label = label
+        self.worktree = worktree
+        self.exit_code = exit_code
+        self.session_id = session_id
 
 
 def _cleanup_variant_worktrees(repo_root: Path, results: list[_VariantResult]) -> None:
@@ -447,17 +312,11 @@ def run_iteration(
 ) -> bool:
     """Run a single iteration of a trigger.
 
-    Args:
-        loop: The trigger (Loop, Subscription, or Schedule) to run
-        iteration: Iteration number
-        run_id: Optional pre-allocated run ID (for scheduler coordination)
-        parent_type: Type prefix for parent ("loop", "subscription", "schedule")
-
+    Works for any trigger type (Loop, Subscription, Schedule).
     Returns True if successful, False on error.
     """
     config = load_config(loop.repo)
 
-    # Create iteration branch from main_branch
     prefix = _iteration_branch_prefix(loop.main_branch)
     branch = f"{prefix}/{iteration:03d}"
     try:
@@ -466,7 +325,6 @@ def run_iteration(
         notify_event("loop.error", {"loop_id": loop.id, "error": f"Failed to create worktree: {e}"})
         return False
 
-    # Create run record
     run = Run(
         id=run_id or str(uuid.uuid4()),
         parent=f"{parent_type}:{loop.id}",
@@ -493,7 +351,6 @@ def run_iteration(
         },
     )
 
-    # Build effective goals (inject adaptive if no mode present)
     effective_goals = build_effective_goals(loop.repo, loop.goals)
     if not effective_goals:
         update_run_status(run.id, RunStatus.FAILED, error="No valid goals found")
@@ -523,7 +380,6 @@ def run_iteration(
         _cleanup_worktree(loop.repo, worktree_path, branch)
         return False
 
-    # Get model configuration
     agent_model = config.agent_model if config else "claude:opus"
     backend, model_variant = parse_model(agent_model)
 
@@ -670,16 +526,13 @@ def run_iteration(
 
         i += 1
 
-    # Clear current step
     update_run_step(run.id, None)
 
-    # Create PR to main_branch and auto-merge (always)
     pr_url = _create_pr_to_main_branch(loop, worktree_path, branch, iteration)
     if pr_url:
         update_run_pr(run.id, pr_url)
         _auto_merge_pr(worktree_path)
 
-        # If LAND mode, also merge main_branch to main
         if loop.merge_mode.value == "land":
             _land_to_main(loop)
 
@@ -697,7 +550,6 @@ def run_iteration(
         },
     )
 
-    # Cleanup worktree
     _cleanup_worktree(loop.repo, worktree_path, branch)
 
     return True
@@ -707,7 +559,6 @@ def _create_pr_to_main_branch(
     loop: Loop, worktree_path: Path, branch: str, iteration: int
 ) -> str | None:
     """Push branch and create PR targeting main_branch."""
-    # Push the branch
     result = subprocess.run(
         ["git", "push", "-u", "origin", branch],
         cwd=worktree_path,
@@ -717,7 +568,6 @@ def _create_pr_to_main_branch(
     if result.returncode != 0:
         return None
 
-    # Generate PR message
     try:
         message = generate_pr_message(worktree_path)
         title = f"[{loop.area_slug}] {message.title}"
@@ -734,7 +584,6 @@ def _create_pr_to_main_branch(
             f"Iteration: {iteration}"
         )
 
-    # Create PR
     cmd = [
         "gh",
         "pr",
@@ -765,18 +614,11 @@ def _auto_merge_pr(worktree_path: Path) -> bool:
 
 
 def _land_to_main(loop: Loop) -> str | None:
-    """Create or update PR from main_branch → main, enable auto-merge.
-
-    Returns PR URL on success, None on failure.
-    Works from main repo (not worktree, which gets deleted).
-    Idempotent: existing PR just gets auto-merge re-enabled.
-    """
+    """Create or update PR from main_branch to main, enable auto-merge."""
     repo = loop.repo
 
-    # Push main_branch
     subprocess.run(["git", "push", "origin", loop.main_branch], cwd=repo, capture_output=True)
 
-    # Check for existing PR
     result = subprocess.run(
         [
             "gh",
@@ -798,7 +640,6 @@ def _land_to_main(loop: Loop) -> str | None:
     existing = json.loads(result.stdout) if result.returncode == 0 and result.stdout.strip() else []
 
     if existing:
-        # PR exists - ensure auto-merge is enabled
         pr_number = existing[0]["number"]
         subprocess.run(
             ["gh", "pr", "merge", str(pr_number), "--squash", "--auto"],
@@ -807,7 +648,6 @@ def _land_to_main(loop: Loop) -> str | None:
         )
         return existing[0]["url"]
 
-    # Create new PR
     result = subprocess.run(
         [
             "gh",
@@ -831,7 +671,6 @@ def _land_to_main(loop: Loop) -> str | None:
 
     pr_url = result.stdout.strip()
 
-    # Enable auto-merge on the new PR
     subprocess.run(["gh", "pr", "merge", "--squash", "--auto"], cwd=repo, capture_output=True)
 
     return pr_url
@@ -844,29 +683,8 @@ def _cleanup_worktree(repo: Path, worktree_path: Path, branch: str) -> None:
     except Exception:
         pass
 
-    # Delete remote branch
     subprocess.run(
         ["git", "push", "origin", "--delete", branch],
         cwd=repo,
         capture_output=True,
     )
-
-
-def main() -> None:
-    """Entry point for background loop runner."""
-    if len(sys.argv) != 2:
-        print("Usage: python -m loopflow.lfd.loop_runner <loop_id>", file=sys.stderr)
-        sys.exit(1)
-
-    loop_id = sys.argv[1]
-    loop = get_loop(loop_id)
-
-    if not loop:
-        print(f"Loop not found: {loop_id}", file=sys.stderr)
-        sys.exit(1)
-
-    run_loop_iterations(loop)
-
-
-if __name__ == "__main__":
-    main()
