+++ src/loopflow/lfd/runs/__init__.py
@@ -1,0 +1,7 @@
+"""Run-centric domain for lfd.
+
+This package contains the core entities that represent work:
+- Run: an execution instance
+- Loop, Subscription, Schedule: parent types that spawn runs
+- Session: legacy tracking for interactive sessions
+"""
